+++ WMF Framework/UILabel+HasText.swift
@@ -1,0 +1,5 @@
+extension UILabel {
+    public var wmf_hasText: Bool {
+        return (text as NSString?)?.length ?? 0 > 0
+    }
+}
